Created TagGeneratorInstructions Component
